Add Iceberg fields to compilation logic (#2009)
* Add Apache Iceberg fields to protos

This change adds fields that will br used to create Apache Iceberg
BigQuery tables to relevant protos.

connection, table_format, file_format and storage_uri are added to
core.proto.

connection, table_format, file_format, storage_uri, bucket_Name,
table_folder_root and table_folder_subpath are added to dataform.proto.

* Add Iceberg options to compile logic

This change ensures Iceberg options provided in the user's config file
are correctly compiled.

* Set defaults for storageUri, remove it from config

This change removes storage_uri as a field users can set in their config
files.

It also sets defaults for tableFolderRoot and tableFolderSubpath. If
tableFolderRoot is not provided to users, we default to `_dataform`. If
tableFolderSubpath is not provided to users, we default to
`{dataset name}/{table name}`.

* Add check for valid file format

This change ensures that the file_format in the user's config file is
set to "PARQUET" because that is the only value currently supported.

* Ensure bucket_name is defined in iceberg config block

This change adds a check for bucket_name in the iceberg subblock of the
user's config file. bucket_name is required when defining iceberg
options.

* Alphabetize imports and use T[] instead of Array<T>

This fixes the build failures.

* Add test for an empty file_format field

* Move file_format to iceberg subblock

This change moves the file_format config field from the bigquery
subblock to the iceberg subblock.

* Move FileFormat enum to IcebergTableConfig

This change moves the FileFormat enum from TableConfig to
IcebergTableConfig because that is where the enum is used.

* Rename Iceberg enums in core.proto

The new naming is consistent with the naming in configs.proto.

* Change fileFormat defaulting logic

We now only default to Iceberg if fileFormat is not defined.

* Add tests for storageUri and connection validation

This change adds a utils_test.ts to hold tests for methods inside
core/utils.ts, and adds tests for methods that validate storageUri and
connection format.

* Remove contextableIcebergOpts

This change removes the temporary state variable contextableIcebergOpts
and sets Iceberg options directly.

* Refactor storageUri construction

* Throw regular instead of session errors

It is not necessary to throe session errors because regular errors get
wrapped into session errors a few layers above.

* Add tests for getters in utils

This change introduces testing for Iceberg-related getters in utils.ts

* Lint fixes

* Refactor Iceberg table tests

This change refactors the Iceberg table tests to check the entire
bigquery object and adds a test with non-Iceberg and Iceberg bigquery
options.

* Add Iceberg options to IncrementalTableConfig

* Add Iceberg options to incremental table compilation

This change adds Iceberg options to the compilation logic for
incremental tables. Unit tests are added to verify the changes.

* Remove validation of iceberg in bigqueryoptions

This change removes the validation of iceberg as part of the
bigqueryoptions in the unverified config. This validation is not
necessary because iceberg is not part of the bigqueryoptions.

* Use existing method to set iceberg proto fields

This change removes the dedicated method for setting iceberg fields and
reuses the existing method for bigquery options.