Enable setting Iceberg config in init CLI (#2026)
* Enable setting Iceberg config in init cli

This change enables users to invoke `dataform init` with an `--iceberg`
option. The users will then be prompted to input defaultBucketName,
defaultTableFolderRoot and defaultTableFolderSubpath. The fields users
input will be set inside workflow_settings.yaml.

* Use map instead of array to store CLI test data

This change switches from an array to a map to store questions and
answers for CLI tests.

* Add CLI input validation for iceberg config

This change adds validation for defaultBucketName,
defaultTableFolderRoot and defaultTableFolderSubpath when reading them
from command line.

* Extend testing for --iceberg CLI flag

Considering we have new checks for input field validity, testing is
extended to cover inputting empty strings.

* Make Iceberg CLI prompts more verbose

* Verify bucket name does not contain 'google'

'google' and close misspellings are prohibited as per
https://cloud.google.com/storage/docs/buckets#naming

* Simplify CLI testing helper functions and states

* Minor fixes