Compile extension interface (#2037)
Introducing a compilation extension interface - the way to plug ad-hoc compilation stages as NPM package without a need to create an explicit JS definition, unblocking the transpilation of additional syntax and formats.

The interface at the moment includes a compilation callback, which receives the original request and Session to operate on.

Two modes are supported:
1. "prologue" - extension performs compilation in session before the regular Dataform compilation of sources is done.
2. "application code" - extension replaces the normal Dataform sources compilation.